refactor: bring scanner, store, and store_gc under the 1000-line cap
CLAUDE.md documents a hard 1000-line-per-file cap enforced by a poly check named
rust-max-lines. That check is not actually in poly.toml, so the cap has never been
enforced and files drifted past it. Split the three worst along real seams, using
the shape the project already prefers (tools.rs -> tools_<area>.rs + helpers*.rs).

  scanner.rs   1274 -> 663   + scanner_file.rs (644)
  store.rs     1117 -> 701   + store_layout.rs (181), blob accessors -> store_blob.rs
  store_gc.rs  1127 -> 784   + store_cache_admin.rs (377)

Seams, not line numbers:
- scanner_file.rs owns one file's journey (process_file, the unchanged-file gates,
  the readers/classifiers, the rayon pool); scanner.rs keeps orchestration
  (candidate enumeration, stale purge, apply_outcomes, post-barrier lanes).
- store_layout.rs answers "which directory holds what" — it changes when the cache
  layout changes, which the Store handle and its index do not.
- store_cache_admin.rs is literally responsibility (2) from store_gc.rs's own module
  doc ("whole-component cleanup and introspection"), disjoint from mark-and-sweep:
  nothing in it reference-counts or reclaims a blob.

Behaviour-preserving by construction: every added line is a re-export, module doc,
import, or visibility widening. Public paths are stable via pub use in each parent,
and the moved Store methods stay inherent, so no call site changed. Proof is the
test count, identical before and after on both feature sets: 839 default, 932 comms,
0 failures.

rust-max-lines stays out of poly.toml for now — four files still exceed the cap and
are being actively edited on another branch of this work.
